+++ CHANGES.rst
@@ -1,3 +1,20 @@
+0.13.0 (2020-03-26)
+-------------------
+New Features
+^^^^^^^^^^^^
+
+- Added two new transforms - ``SphericalToCartesian`` and
+  ``CartesianToSpherical``. [#275, #284, #285]
+
+- Added ``to_fits_sip`` method to generate FITS header with SIP keywords [#286]
+
+- Added ``get_ctype_from_ucd`` function. [#288]
+
+Bug Fixes
+^^^^^^^^^
+
+- Fixed an off by one issue in ``utils.make_fitswcs_transform``. [#290]
+
 0.12.0 (2019-12-24)
 -------------------
 New Features